+++ Source/JavaScriptCore/assembler/ARM64EAssembler.h
@@ -284,72 +284,6 @@ public:
     ALWAYS_INLINE void eretaa() { insn(encodeGroup4(Group4Op::ERETAA)); }
     ALWAYS_INLINE void eretab() { insn(encodeGroup4(Group4Op::ERETAB)); }
 
-    enum ExoticAtomicLoadStoreOp {
-        ExoticAtomicLoadStoreOp_Add   = 0b0'000'00,
-        ExoticAtomicLoadStoreOp_Clear = 0b0'001'00,
-        ExoticAtomicLoadStoreOp_Xor   = 0b0'010'00,
-        ExoticAtomicLoadStoreOp_Set   = 0b0'011'00,
-        ExoticAtomicLoadStoreOp_Swap  = 0b1'000'00,
-    };
-
-    static int exoticAtomicLoadStore(MemOpSize size, ExoticAtomicLoadStoreOp op, ExoticLoadFence loadFence, ExoticStoreFence storeFence, RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        ASSERT((rs & 0b11111) == rs);
-        ASSERT((rn & 0b11111) == rn);
-        ASSERT((rt & 0b11111) == rt);
-        return 0b00111000'00100000'00000000'00000000 | size << 30 | loadFence << 23 | storeFence << 22 | rs << 16 | op << 10 | rn << 5 | rt;
-    }
-
-    static int exoticAtomicCAS(MemOpSize size, ExoticLoadFence loadFence, ExoticStoreFence storeFence, RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        ASSERT((rs & 0b11111) == rs);
-        ASSERT((rn & 0b11111) == rn);
-        ASSERT((rt & 0b11111) == rt);
-        return 0b00001000'10100000'01111100'00000000 | size << 30 | storeFence << 22 | rs << 16 | loadFence << 15 | rn << 5 | rt;
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void ldaddal(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicLoadStore(MEMOPSIZE, ExoticAtomicLoadStoreOp_Add, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void ldeoral(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicLoadStore(MEMOPSIZE, ExoticAtomicLoadStoreOp_Xor, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void ldclral(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicLoadStore(MEMOPSIZE, ExoticAtomicLoadStoreOp_Clear, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void ldsetal(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicLoadStore(MEMOPSIZE, ExoticAtomicLoadStoreOp_Set, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void swpal(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicLoadStore(MEMOPSIZE, ExoticAtomicLoadStoreOp_Swap, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
-    template<int datasize>
-    ALWAYS_INLINE void casal(RegisterID rs, RegisterID rt, RegisterID rn)
-    {
-        CHECK_MEMOPSIZE();
-        insn(exoticAtomicCAS(MEMOPSIZE, ExoticLoadFence_Acquire, ExoticStoreFence_Release, rs, rt, rn));
-    }
-
     // Overload of the ARM64 equivalents.
 
     // Needed because we need to call our overloaded linkPointer below.
