+++ Source/JavaScriptCore/assembler/MacroAssemblerARM64E.h
@@ -372,126 +372,6 @@ public:
 #endif
             m_assembler.retab();
     }
-
-    void atomicXchgAdd8(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldaddal<8>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgAdd16(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldaddal<16>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgAdd32(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldaddal<32>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgAdd64(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldaddal<64>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgXor8(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldeoral<8>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgXor16(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldeoral<16>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgXor32(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldeoral<32>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgXor64(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldeoral<64>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgOr8(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldsetal<8>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgOr16(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldsetal<16>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgOr32(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldsetal<32>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgOr64(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldsetal<64>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgClear8(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldclral<8>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgClear16(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldclral<16>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgClear32(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldclral<32>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchgClear64(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.ldclral<64>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchg8(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.swpal<8>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchg16(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.swpal<16>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchg32(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.swpal<32>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicXchg64(RegisterID src, Address address, RegisterID dest)
-    {
-        m_assembler.swpal<64>(src, dest, extractSimpleAddress(address));
-    }
-
-    void atomicStrongCAS8(RegisterID expectedAndResult, RegisterID newValue, Address address)
-    {
-        m_assembler.casal<8>(expectedAndResult, newValue, extractSimpleAddress(address));
-    }
-
-    void atomicStrongCAS16(RegisterID expectedAndResult, RegisterID newValue, Address address)
-    {
-        m_assembler.casal<16>(expectedAndResult, newValue, extractSimpleAddress(address));
-    }
-
-    void atomicStrongCAS32(RegisterID expectedAndResult, RegisterID newValue, Address address)
-    {
-        m_assembler.casal<32>(expectedAndResult, newValue, extractSimpleAddress(address));
-    }
-
-    void atomicStrongCAS64(RegisterID expectedAndResult, RegisterID newValue, Address address)
-    {
-        m_assembler.casal<64>(expectedAndResult, newValue, extractSimpleAddress(address));
-    }
 };
 
 } // namespace JSC
