+++ Source/JavaScriptCore/b3/air/AirOpcode.opcodes
@@ -217,7 +217,7 @@ arm64: Sub64 U:G:64, U:G:64, D:G:64
     Tmp, Tmp, Tmp, Tmp, Tmp, Tmp
 
 SubDouble U:F:64, U:F:64, D:F:64
-    arm: Tmp, Tmp, Tmp
+    arm x86_64_avx: Tmp, Tmp, Tmp
     x86: Tmp, Addr, Tmp
     x86: Tmp, Index, Tmp
 
@@ -226,7 +226,7 @@ x86: SubDouble U:F:64, UD:F:64
     Addr, Tmp
 
 SubFloat U:F:32, U:F:32, D:F:32
-    arm: Tmp, Tmp, Tmp
+    arm x86_64_avx: Tmp, Tmp, Tmp
     x86: Tmp, Addr, Tmp
     x86: Tmp, Index, Tmp
 
@@ -343,14 +343,14 @@ x86: MulFloat U:F:32, UD:F:32
     Tmp, Tmp
     Addr, Tmp
 
-arm: DivDouble U:F:64, U:F:64, D:F:64
+arm x86_64_avx: DivDouble U:F:64, U:F:64, D:F:64
     Tmp, Tmp, Tmp
 
 x86: DivDouble U:F:64, UD:F:64
     Tmp, Tmp
     Addr, Tmp
 
-arm: DivFloat U:F:32, U:F:32, D:F:32
+arm x86_64_avx: DivFloat U:F:32, U:F:32, D:F:32
     Tmp, Tmp, Tmp
 
 x86: DivFloat U:F:32, UD:F:32
@@ -1097,25 +1097,25 @@ x86_64: AtomicStrongCAS64 U:G:32, UD:G:64, U:G:64, UD:G:64, ZD:G:8 /effects
     StatusCond, Tmp*, Tmp, Addr, Tmp
     StatusCond, Tmp*, Tmp, Index, Tmp
 
-x86 arm64e: AtomicStrongCAS8 UD:G:8, U:G:8, UD:G:8 /effects
+x86 arm64_lse: AtomicStrongCAS8 UD:G:8, U:G:8, UD:G:8 /effects
     x86: Tmp*, Tmp, Addr
     x86: Tmp*, Tmp, Index
-    arm64e: Tmp, Tmp, SimpleAddr
+    arm64_lse: Tmp, Tmp, SimpleAddr
 
-x86 arm64e: AtomicStrongCAS16 UD:G:16, U:G:32, UD:G:16 /effects
+x86 arm64_lse: AtomicStrongCAS16 UD:G:16, U:G:32, UD:G:16 /effects
     x86: Tmp*, Tmp, Addr
     x86: Tmp*, Tmp, Index
-    arm64e: Tmp, Tmp, SimpleAddr
+    arm64_lse: Tmp, Tmp, SimpleAddr
 
-x86 arm64e: AtomicStrongCAS32 UD:G:32, U:G:32, UD:G:32 /effects
+x86 arm64_lse: AtomicStrongCAS32 UD:G:32, U:G:32, UD:G:32 /effects
     x86: Tmp*, Tmp, Addr
     x86: Tmp*, Tmp, Index
-    arm64e: Tmp, Tmp, SimpleAddr
+    arm64_lse: Tmp, Tmp, SimpleAddr
 
-x86_64 arm64e: AtomicStrongCAS64 UD:G:64, U:G:64, UD:G:64 /effects
+x86_64 arm64_lse: AtomicStrongCAS64 UD:G:64, U:G:64, UD:G:64 /effects
     x86_64: Tmp*, Tmp, Addr
     x86_64: Tmp*, Tmp, Index
-    arm64e: Tmp, Tmp, SimpleAddr
+    arm64_lse: Tmp, Tmp, SimpleAddr
 
 x86: BranchAtomicStrongCAS8 U:G:32, UD:G:8, U:G:8, UD:G:8 /branch /effects
     StatusCond, Tmp*, Tmp, Addr
@@ -1378,64 +1378,64 @@ arm64: Depend32 U:G:32, ZD:G:32
 arm64: Depend64 U:G:64, ZD:G:64
     Tmp, Tmp
 
-arm64e: AtomicXchgAdd8 U:G:8, UD:G:8, ZD:G:32 /effects
+arm64_lse: AtomicXchgAdd8 U:G:8, UD:G:8, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgAdd16 U:G:16, UD:G:16, ZD:G:32 /effects
+arm64_lse: AtomicXchgAdd16 U:G:16, UD:G:16, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgAdd32 U:G:32, UD:G:32, ZD:G:32 /effects
+arm64_lse: AtomicXchgAdd32 U:G:32, UD:G:32, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgAdd64 U:G:64, UD:G:64, ZD:G:64 /effects
+arm64_lse: AtomicXchgAdd64 U:G:64, UD:G:64, ZD:G:64 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgOr8 U:G:8, UD:G:8, ZD:G:32 /effects
+arm64_lse: AtomicXchgOr8 U:G:8, UD:G:8, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgOr16 U:G:16, UD:G:16, ZD:G:32 /effects
+arm64_lse: AtomicXchgOr16 U:G:16, UD:G:16, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgOr32 U:G:32, UD:G:32, ZD:G:32 /effects
+arm64_lse: AtomicXchgOr32 U:G:32, UD:G:32, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgOr64 U:G:64, UD:G:64, ZD:G:64 /effects
+arm64_lse: AtomicXchgOr64 U:G:64, UD:G:64, ZD:G:64 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgClear8 U:G:8, UD:G:8, ZD:G:32 /effects
+arm64_lse: AtomicXchgClear8 U:G:8, UD:G:8, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgClear16 U:G:16, UD:G:16, ZD:G:32 /effects
+arm64_lse: AtomicXchgClear16 U:G:16, UD:G:16, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgClear32 U:G:32, UD:G:32, ZD:G:32 /effects
+arm64_lse: AtomicXchgClear32 U:G:32, UD:G:32, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgClear64 U:G:64, UD:G:64, ZD:G:64 /effects
+arm64_lse: AtomicXchgClear64 U:G:64, UD:G:64, ZD:G:64 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgXor8 U:G:8, UD:G:8, ZD:G:32 /effects
+arm64_lse: AtomicXchgXor8 U:G:8, UD:G:8, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgXor16 U:G:16, UD:G:16, ZD:G:32 /effects
+arm64_lse: AtomicXchgXor16 U:G:16, UD:G:16, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgXor32 U:G:32, UD:G:32, ZD:G:32 /effects
+arm64_lse: AtomicXchgXor32 U:G:32, UD:G:32, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchgXor64 U:G:64, UD:G:64, ZD:G:64 /effects
+arm64_lse: AtomicXchgXor64 U:G:64, UD:G:64, ZD:G:64 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchg8 U:G:8, UD:G:8, ZD:G:32 /effects
+arm64_lse: AtomicXchg8 U:G:8, UD:G:8, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchg16 U:G:16, UD:G:16, ZD:G:32 /effects
+arm64_lse: AtomicXchg16 U:G:16, UD:G:16, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchg32 U:G:32, UD:G:32, ZD:G:32 /effects
+arm64_lse: AtomicXchg32 U:G:32, UD:G:32, ZD:G:32 /effects
     Tmp, SimpleAddr, Tmp
 
-arm64e: AtomicXchg64 U:G:64, UD:G:64, ZD:G:64 /effects
+arm64_lse: AtomicXchg64 U:G:64, UD:G:64, ZD:G:64 /effects
     Tmp, SimpleAddr, Tmp
 
 Compare32 U:G:32, U:G:32, U:G:32, ZD:G:32
