+++ Source/WTF/wtf/PlatformHave.h
@@ -263,6 +263,10 @@
 #define HAVE_FJCVTZS_INSTRUCTION 1
 #endif
 
+#if (CPU(ARM64E) && OS(DARWIN)) || (COMPILER(CLANG) && defined(__ARM_FEATURE_ATOMICS))
+#define HAVE_LSE_INSTRUCTION 1
+#endif
+
 #if PLATFORM(COCOA) && !PLATFORM(WATCHOS) && !PLATFORM(APPLETV) && !PLATFORM(MACCATALYST)
 #define HAVE_APP_LINKS 1
 #endif
